test: update test for ChangeWalletPasswordDialog and BackupSecretPhraseDialog (#1)
